Add multiple dns enteries if public ips are set.
The automatic dns entries for services use the namespace in the
dns name:

guestbook.default.example.com

If a service has public ips set, it makes sense to use a round
robin and omit the domain name. For example, if there are two public
ips set, there will be two records for the above service:

0.guestbook.example.com
1.guestbook.example.com

guestbook.example.com will be a round robin to both of them.